+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.1
+
+- Fix stress reduces HP to early
+- Fix Evasion was limited to 6 when increasing the value with arrow keys
+- Add Attack Button for Players so Attacks are rolled with a duality roll and adjust Hope/Fear/Stress
+- Add highlight in The Roll Log to the Total if a roll was a duality roll
+
 # 1.0.0
 
 - Initial version of GMD